# openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A simple pastebin service with analytics
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pastebin API
  version: "1.0"
paths:
...
  "/p/{slug}":
    get:
      description: Retrieve a public paste by its URL slug
      parameters:
        - description: Paste slug
          in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paste data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.PasteOutput"
        "400":
          description: Invalid slug
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "404":
          description: Paste not found
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to get paste
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: Get public paste by slug
      tags:
        - pastes
  /paste:
    post:
      description: Create a new paste with optional expiration
      parameters:
        - description: Expiration duration (e.g., '24h', '7d')
          in: query
          name: expires_in
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/models.PasteInput"
        description: Paste data
        required: true
      responses:
        "201":
          description: Created paste with shareable URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.PasteOutput"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to create paste
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Create a new paste
      tags:
        - pastes
  "/paste/{id}":
    delete:
      description: Delete a specific paste by its ID
      parameters:
        - description: Paste ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paste deleted successfully
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "400":
          description: Invalid paste ID
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to delete paste
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Delete paste by ID
      tags:
        - pastes
    get:
      description: Retrieve a specific paste by its ID. Password required for
        password-protected pastes if user is not the owner.
      parameters:
        - description: Paste ID
          in: path
          name: id
          required: true
          schema:
            type: string
        - description: Password for password-protected pastes
          in: query
          name: password
          schema:
            type: string
      responses:
        "200":
          description: Paste data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.PasteOutput"
        "400":
          description: Invalid paste ID or missing password
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to get paste
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: Get paste by ID
      tags:
        - pastes
    put:
      description: Update an existing paste by ID
      parameters:
        - description: Paste ID
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/models.PatchPaste"
        description: Paste update data
        required: true
      responses:
        "200":
          description: Paste updated successfully
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to update paste
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Update a paste
      tags:
        - pastes
  /paste/filter:
    get:
      description: Filter pastes by criteria
      parameters:
        - description: Filter by languages
          in: query
          name: languages
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
        - description: Filter by start date (RFC3339)
          in: query
          name: date_from
          schema:
            type: string
        - description: Filter by end date (RFC3339)
          in: query
          name: date_to
          schema:
            type: string
        - description: Field to sort by
          in: query
          name: sort_by
          schema:
            type: string
        - description: Sort order (asc or desc)
          in: query
          name: sort_order
          schema:
            type: string
      responses:
        "200":
          description: Filtered pastes
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/models.PasteOutput"
                type: array
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to filter pastes
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Filter pastes
      tags:
        - pastes
  /pastes:
    get:
      description: Retrieve all pastes for the authenticated user with pagination
      parameters:
        - description: "Number of pastes to return (default: 10, max: 100)"
          in: query
          name: limit
          schema:
            type: integer
        - description: "Number of pastes to skip (default: 0)"
          in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of pastes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.PaginatedPastesResponse"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to get pastes
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Get all pastes for user
      tags:
        - pastes
...
  "/raw/{slug}":
    get:
      description: Retrieve raw text content of a public paste by its URL slug
      parameters:
        - description: Paste slug
          in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Raw paste content
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Invalid slug
          content:
            text/plain:
              schema:
                additionalProperties:
                  type: string
                type: object
        "404":
          description: Paste not found
          content:
            text/plain:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Unable to get paste
          content:
            text/plain:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: Get raw paste content by slug
      tags:
        - pastes
...
servers:
  - url: //localhost:8080
components:
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
  schemas:
...
    models.PaginatedPastesResponse:
      properties:
        has_more:
          type: boolean
        limit:
          type: integer
        offset:
          type: integer
        pastes:
          items:
            $ref: "#/components/schemas/models.PasteOutput"
          type: array
        total:
          type: integer
      type: object
    models.PasteInput:
      properties:
        content:
          type: string
        expires_at:
          type: string
        expires_in:
          type: string
        language:
          type: string
        password:
          type: string
        title:
          type: string
      type: object
    models.PasteOutput:
      properties:
        content:
          type: string
        created_at:
          type: string
        expires_at:
          type: string
        id:
          type: string
        is_private:
          type: boolean
        language:
          type: string
        title:
          type: string
        updated_at:
          type: string
        url:
          type: string
        user_id:
          type: string
        views:
          type: integer
      type: object
    models.PatchPaste:
      properties:
        content:
          type: string
        expires_at:
          type: string
        id:
          type: string
        is_private:
          type: boolean
        language:
          type: string
        password:
          type: string
        title:
          type: string
        user_id:
          type: string
      type: object
...